com.unity.ai.toolkit@1.6.0-pre.1 ## [1.6.0-pre.1] - 2026-02-09

+++ Modules/Accounts/Services/States/PointsBalanceState.cs
@@ -27,7 +27,7 @@ namespace Unity.AI.Toolkit.Accounts.Services.States
             AIDropdownBridge.UserStateChanged(Refresh);
         }
         
-        public bool CanAfford(int cost) => Value != null && Value.PointsAvailable >= cost;        
+        public bool CanAfford(long cost) => Value != null && Value.PointsAvailable >= cost;        
 
         async Task RefreshInternal()
         {
